At Thoughtful, I've begun to focus on helping s...
At Thoughtful, I've begun to focus on helping small companies run their tech by keeping technology simpler than they thought possible using low-code tools that work out of the box and reducing the number of moving parts in complex systems. Think of me as your on-demand CTO as a service. I keep a small number of long-term clients, most of which are focused on some very interesting business but don't have a core competency in internet technology - for example, there's a company in SF who is working on lab-grown meat that needed a custom database installation for their genetics system. They have a lot of biotech engineers, but not a single software engineer. Another company had built an impressively sophisticated subscription + CRM system but were experiencing growing pains and unable to manage it, so I helped them move to more low-code, scalable cloud platforms and now they are getting close to running the company without any engineers at all. This is the kind of project that I can make the biggest difference in.

Each post, I plan to give you a high-level tidbit that I think you should be aware of as an entrepreneur running a tech-enabled company. This month I’ve been thinking about how the tech industry is moving at light speed and using the right tools to get the most done with the smallest team is the key to success. This is why I believe in low-code tooling. Secondly, as an extension of this thinking, I believe that reducing infrastructure and moving parts makes it possible to build tech-enabled companies that don’t require full-time engineering teams. For the past few years, I have been focused on Docker + Heroku as my preferred backend stack, but I am now beginning to explore Serverless as I believe it requires even less moving parts than Docker.